Updated package name and class names; added separate configurations for client and server
