+++ examples/example-gcp-observability/src/main/resources/io/grpc/examples/gcpObservability/gcp_observability_server_config.json
@@ -12,6 +12,6 @@
     }]
   },
   "labels": {
-    "environment" : "observability-example"
+    "environment" : "example-server"
   }
 }
